rename : OtpDto 파일 디렉토리 변경, ApiModel 어노테이션 추가
