+++ src/main/java/com/grablunchtogether/dto/user/OtpDto.java
@@ -1,6 +1,7 @@
-package com.grablunchtogether.dto;
+package com.grablunchtogether.dto.user;
 
 
+import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,6 +12,7 @@ public class OtpDto {
     @AllArgsConstructor
     @NoArgsConstructor
     @Builder
+    @ApiModel("Otp검증 요청")
     public static class Request {
       private String otp;
       private String email;
